+++ src/server.ts
@@ -40,9 +40,7 @@ import { TradeAggregationCallBuilder } from "./trade_aggregation_call_builder";
 import { TradesCallBuilder } from "./trades_call_builder";
 import { TransactionCallBuilder } from "./transaction_call_builder";
 
-import HorizonAxiosClient, {
-  getCurrentServerTime,
-} from "./horizon_axios_client";
+import AxiosClient, { getCurrentServerTime } from "./axios_client";
 
 export const SUBMIT_TRANSACTION_TIMEOUT = 60 * 1000;
 
@@ -91,7 +89,7 @@ export class Server {
       customHeaders["X-App-Version"] = opts.appVersion;
     }
     if (!isEmpty(customHeaders)) {
-      HorizonAxiosClient.interceptors.request.use((config) => {
+      AxiosClient.interceptors.request.use((config) => {
         // merge the custom headers with an existing headers
         config.headers = merge(customHeaders, config.headers);
 
@@ -138,7 +136,7 @@ export class Server {
     seconds: number,
     _isRetry: boolean = false,
   ): Promise<Server.Timebounds> {
-    // HorizonAxiosClient instead of this.ledgers so we can get at them headers
+    // AxiosClient instead of this.ledgers so we can get at them headers
     const currentTime = getCurrentServerTime(this.serverURL.hostname());
 
     if (currentTime) {
@@ -158,7 +156,7 @@ export class Server {
 
     // otherwise, retry (by calling the root endpoint)
     // toString automatically adds the trailing slash
-    await HorizonAxiosClient.get(URI(this.serverURL as any).toString());
+    await AxiosClient.get(URI(this.serverURL as any).toString());
     return await this.fetchTimebounds(seconds, true);
   }
 
@@ -307,7 +305,7 @@ export class Server {
         .toString("base64"),
     );
 
-    return HorizonAxiosClient.post(
+    return AxiosClient.post(
       URI(this.serverURL as any)
         .segment("transactions")
         .toString(),