+++ test/test-browser.js
@@ -1,4 +1,4 @@
 /* eslint-disable no-undef */
 
 window.axios = StellarSdk.axios;
-window.HorizonAxiosClient = StellarSdk.HorizonAxiosClient;
+window.AxiosClient = StellarSdk.AxiosClient;